rewrite tests to ts
