+++ package.json
@@ -6,6 +6,7 @@
   "jsnext:main": "dist/satellite.es.js",
   "types": "types/index.d.ts",
   "module": "dist/satellite.es.js",
+  "type": "module",
   "scripts": {
     "build": "rimraf lib dist compiled types && tsc && npm run transpile && npm run dist",
     "transpile": "babel compiled --out-dir lib --ignore indexUmd.js && node --require @babel/register commands/add-copyright.js 'lib/**/*.js'",
@@ -18,9 +19,9 @@
     "copy": "copyfiles -u 1 'dist/**/*' sgp4_verification/lib/sgp4",
     "lint": "eslint src",
     "lint:test": "eslint test",
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "test:vitest": "vitest run",
+    "test": "vitest run",
+    "test:watch": "vitest watch",
+    "test:coverage": "vitest run --coverage",
     "raise": "raise-version",
     "prepublishOnly": "npm run lint && npm run lint:test && npm test && npm run build",
     "wasm:serve": "vite serve simd-test",
@@ -61,6 +62,7 @@
     "@eslint/js": "^9.22.0",
     "@rollup/plugin-babel": "^6.0.3",
     "@rollup/plugin-terser": "^0.2.1",
+    "@vitest/coverage-v8": "^3.2.4",
     "chokidar": "^3.5.3",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
