+++ src/indexUmd.ts
@@ -6,7 +6,7 @@ import { propagate, sgp4, gstime } from './propagation.js';
 import { SatRecError } from './propagation/SatRec.js';
 import * as types from './common-types.js';
 
-import dopplerFactor from './dopplerFactor.js';
+import { dopplerFactor } from './dopplerFactor.js';
 
 import {
   radiansToDegrees,
