--- test/propagation/dsinit.test.js
@@ -1,11 +1,0 @@
-import dsinit from '../../compiled/propagation/dsinit';
-import dsOptionSets from './dsinit.json';
-
-describe('Initialize Resonance Terms', () => {
-  it('Geopotential Resonance for 12 Hour Orbits', () => {
-    dsOptionSets.forEach((testSet) => {
-      const results = dsinit(testSet.options);
-      expect(results).toEqual(testSet.results);
-    });
-  });
-});
