--- test/transforms.test.js
@@ -1,104 +1,0 @@
-import {
-  degreesLat,
-  degreesLong,
-  radiansLat,
-  radiansLong,
-  geodeticToEcf,
-  eciToGeodetic,
-  eciToEcf,
-  ecfToEci,
-  ecfToLookAngles,
-} from '../dist/satellite';
-import compareVectors from './compareVectors';
-import transformData from './transforms.json';
-
-const numDigits = 6;
-
-describe('Latitude & longitude conversions', () => {
-  const {
-    validLatitudes,
-    validLongitudes,
-    validGeodeticToEcf,
-    validEciToGeodetic,
-    validEciToEcf,
-    validEcfToEci,
-    validEcfToLookangles,
-    invalidLatitudes,
-    invalidLongitudes,
-  } = transformData;
-
-  validLatitudes.forEach((item) => {
-    it(`convert valid latitude value (${item.radians} radians) to degrees`, () => {
-      expect(degreesLat(item.radians)).toBeCloseTo(item.degrees, numDigits);
-    });
-    it(`convert valid latitude value (${item.degrees} degrees) to radians`, () => {
-      expect(radiansLat(item.degrees)).toBeCloseTo(item.radians, numDigits);
-    });
-  });
-
-  validLongitudes.forEach((item) => {
-    it(`convert valid longitude value (${item.radians} radians) to degrees`, () => {
-      expect(degreesLong(item.radians)).toBeCloseTo(item.degrees, numDigits);
-    });
-    it(`convert valid longitude value (${item.degrees} degrees) to radians`, () => {
-      expect(radiansLong(item.degrees)).toBeCloseTo(item.radians, numDigits);
-    });
-  });
-
-  validGeodeticToEcf.forEach((item) => {
-    it('convert valid LLA coordinates to ECF', () => {
-      const ecfCoordinates = geodeticToEcf(item.lla);
-      compareVectors(ecfCoordinates, item.ecf, 8);
-    });
-  });
-
-  validEciToGeodetic.forEach((item) => {
-    it('convert valid ECI coordinates to LLA', () => {
-      const llaCoordinates = eciToGeodetic(item.eci, item.gmst);
-      expect(llaCoordinates.longitude).toBeCloseTo(item.lla.longitude);
-      expect(llaCoordinates.latitude).toBeCloseTo(item.lla.latitude);
-      expect(llaCoordinates.height).toBeCloseTo(item.lla.height);
-    });
-  });
-
-  validEciToEcf.forEach((item) => {
-    it('convert valid ECI coordinates to ECF', () => {
-      const ecfCoordinates = eciToEcf(item.eci, item.gmst);
-      compareVectors(ecfCoordinates, item.ecf, 8);
-    });
-  });
-
-  validEcfToEci.forEach((item) => {
-    it('convert valid ECF coordinates to ECI', () => {
-      const eciCoordinates = ecfToEci(item.ecf, item.gmst);
-      compareVectors(eciCoordinates, item.eci, 8);
-    });
-  });
-
-  validEcfToLookangles.forEach((item) => {
-    it('convert valid ECF coordinates to RAE', () => {
-      const raeCoordinates = ecfToLookAngles(item.lla, item.satelliteEcf);
-      expect(raeCoordinates.rangeSat).toBeCloseTo(item.rae.rangeSat);
-      expect(raeCoordinates.azimuth).toBeCloseTo(item.rae.azimuth);
-      expect(raeCoordinates.elevation).toBeCloseTo(item.rae.elevation);
-    });
-  });
-
-  invalidLatitudes.forEach((item) => {
-    it(`convert invalid latitude value (${item.radians} radians) to degrees`, () => {
-      expect(() => degreesLat(item.radians)).toThrowError(RangeError);
-    });
-    it(`convert invalid latitude value (${item.degrees} degrees) to radians`, () => {
-      expect(() => radiansLat(item.degrees)).toThrowError(RangeError);
-    });
-  });
-
-  invalidLongitudes.forEach((item) => {
-    it(`convert invalid longitude value (${item.radians} radians) to degrees`, () => {
-      expect(() => degreesLong(item.radians)).toThrowError(RangeError);
-    });
-    it(`convert invalid longitude value (${item.degrees} degrees) to radians`, () => {
-      expect(() => radiansLong(item.degrees)).toThrowError(RangeError);
-    });
-  });
-});
