+++ vitest.config.ts
@@ -2,7 +2,9 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    root: './src/wasm-wrapping',
+    include: [
+      'test/**/*.test.ts',
+    ],
     environment: 'node',
   }
 })
